Create native_alg
The number 1 is not considered simple by definition, this case can be handled separately.
For the remaining numbers, we will perform a check: if the number nn is divisible without remainder by some number ii, such that 1 < i < n1< i < n, then nn is composite.
We need to test separately that the number 2 will be considered prime. This is an important special case, because there are no integers between 1 and 2, which means that the divisibility check will never be performed.